Module Hash.Noalloc

Versions of these functions which write their output in a buffer passed in as an argument

Direct interface

val hash : alg:SharedDefs.HashDefs.alg -> msg:bytes -> digest:bytes -> unit

hash alg msg digest hashes msg using algorithm alg and outputs the result in digest.

Streaming interface

val finish : st:t -> digest:bytes -> unit

finish st digest writes a digest in digest, without invalidating the internal state st.