Module HMAC_SHA2_384.Noalloc

Version of this function which writes its output in a buffer passed in as an argument

val mac : key:bytes -> msg:bytes -> tag:bytes -> unit

mac key msg tag computes the MAC of msg using key key and writes the result in tag. The `tag` buffer needs to satisfy the size requirements for the output buffer.