Module Keccak.Noalloc

Versions of these functions which write their output in a buffer passed in as an argument

val shake128 : msg:bytes -> digest:bytes -> unit

shake128 msg size hashes msg using SHAKE-128 and returns a digest of size bytes.

val shake256 : msg:bytes -> digest:bytes -> unit

shake256 msg digest hashes msg using SHAKE-256 and outputs the result in digest.

val keccak : rate:int -> capacity:int -> suffix:int -> msg:bytes -> digest:bytes -> unit

Direct access to the general Keccak function, of which all the SHA-3 and SHAKE functions are instances. While the library does run some sanity checks for the parameters, users should be extremely careful if using the Keccak function directly.