Module Hacl.MD5

Direct hashing with MD5

This function should not be used for cryptographic applications!

The digest buffer must match the digest size of MD5, which is 16 bytes.

type bytes = SharedDefs.CBytes.t

val hash : bytes -> bytes

hash msg returns the hash of msg.

val hash_noalloc : msg:bytes -> digest:bytes -> unit

hash_noalloc msg digest hashes msg and outputs the result in digest.