Module SHA2_256.Noalloc

Versions of these functions which write their output in a buffer passed in as an argument

val sign : sk:bytes -> msg:bytes -> k:bytes -> signature:bytes -> bool

sign sk msg k signature attempts to sign the message msg with secret key sk and signing secret k. If successful, the signature is written in signature and the function returns true.