Module Hacl.RandomBuffer

A randomness function implemented with platform-dependent code for Unix and Windows

The randombytes function is handwritten, unverified C code. In Unix, it is implemented using the getrandom syscall, with a fallback to /dev/urandom. In Windows, it is implemented using CryptGenRandom.

val randombytes : size:int -> bytes option

randombytes size attempts to create a buffer containing size random bytes

val randombytes_noalloc : out:bytes -> bool

randombytes_noalloc out attempts to fill out with random bytes and returns true if successful.