Module Hacl.SHA3_256

Direct hashing with SHA3-256

The digest buffer must match the digest size of SHA3-256, which is 32 bytes.

type bytes = SharedDefs.CBytes.t

val hash : bytes -> bytes

hash msg returns the hash of msg.

val hash_noalloc : msg:bytes -> digest:bytes -> unit

hash_noalloc msg digest hashes msg and outputs the result in digest.