Module Hacl.SHA3_512

Direct hashing with SHA3-512

The digest buffer must match the digest size of SHA3-512, which is 64 bytes.

type bytes = SharedDefs.CBytes.t

val hash : bytes -> bytes

hash msg returns the hash of msg.

module Noalloc : sig ... end

Version of this function which writes its output in a buffer passed in as an argument