Module DRBG.Noalloc

Versions of these functions which write their output in a buffer passed in as an argument

val generate : ?⁠additional_input:bytes -> t -> bytes -> bool

generate ?additional_input st output takes an optional additional_input, a state st and an output buffer output, which will be filled with random bytes if successful.