Module Hacl_star__EverCrypt.HKDF

Agile, multiplexing interface for HKDF

Supports the same hashing algorithms as EverCrypt.HMAC.

val extract : alg:Hacl_star.SharedDefs.HashDefs.alg -> salt:bytes -> ikm:bytes -> bytes

extract alg salt ikm computes a pseudorandom key using hashing algorithm alg with input key material ikm and salt salt.

val expand : alg:Hacl_star.SharedDefs.HashDefs.alg -> prk:bytes -> info:bytes -> size:int -> bytes

expand alg prk info size expands the pseudorandom key prk using hashing algorithm alg, taking the info string info into account and returns a buffer of size bytes.

module Noalloc : sig ... end

Versions of these functions which write their output in a buffer passed in as an argument