Module HKDF_SHA2_384.Noalloc

Versions of these functions which write their output in a buffer passed in as an argument

val extract : salt:bytes -> ikm:bytes -> prk:bytes -> unit

extract salt ikm prk computes a pseudorandom key prk using input key material ikm and salt salt.

val expand : prk:bytes -> info:bytes -> okm:bytes -> unit

expand prk info okm expands the pseudorandom key prk, taking the info string info into account, and writes the output key material in okm.