Module Hacl_star__Hacl.NaCl

Box (public-key authenticated encryption) and Secretbox (secret-key authenticated encryption)

Portable C implementations offering both the easy and detached interfaces of Box and Secretbox. For Box, the precomputation interface is also supported.

val box_beforenm : pk:bytes -> sk:bytes -> bytes option

box_beforenm pk sk precomputes a 32-byte X25519 shared key ck using one party's 32-byte public key pk and the other party's 32-byte secret key sk. The shared key can then be used in the Box precomputation interface (box_afternm and box_open_afternm functions) in both Easy and Detached.

This is useful when calling the functions repeatedly, as it avoids computing the shared key on every function call.

val box_beforenm_noalloc : pk:bytes -> sk:bytes -> ck:bytes -> bool

box_beforenm_noalloc pk sk ck is a version of box_beforenm which takes an additional argument ck where the result is written, returning `true` if it is successful, similar to the functions in NaCl.Easy.Noalloc and NaCl.Detached.

Buffers pk, sk, and ck must be distinct.

module Easy : sig ... end

The easy interface concatenates the ciphertext and the 16-byte long message authentication tag into a single buffer.

module Detached : sig ... end

The detached interface uses 2 separate buffers for the ciphertext and the message authentication tag. This allows users to encrypt and decrypt data in-place, by passing the same buffer for both plaintext and ciphertext.