Module P256.SHA2_512

type bytes = Hacl_star.SharedDefs.CBytes.t

val sign : sk:bytes -> msg:bytes -> k:bytes -> bytes option

sign sk msg k attempts to sign the message msg with secret key sk and signing secret k and returns the signature if successful.

val sign_noalloc : sk:bytes -> msg:bytes -> k:bytes -> signature:bytes -> bool

sign_noalloc sk msg k signature attempts to sign the message msg with secret key sk and signing secret k. If successful, the signature is written in signature and the function returns true.

val verify : pk:bytes -> msg:bytes -> signature:bytes -> bool

verify pk msg signature checks the signature of msg using public key pk and returns true if it is valid.