Module RandomBuffer.Noalloc

Version of this function which writes its output in a buffer passed in as an argument

val randombytes : out:bytes -> bool

randombytes out attempts to fill out with random bytes and returns true if successful.