Module Hacl_star__Hacl.SHA3_384

Direct hashing with SHA3-384

The digest buffer must match the digest size of SHA3-384, which is 48 bytes.

type bytes = Hacl_star.SharedDefs.CBytes.t

val hash : bytes -> bytes

hash msg returns the hash of msg.

val hash_noalloc : msg:bytes -> digest:bytes -> unit

hash_noalloc msg digest hashes msg and outputs the result in digest.