Module type Hacl_star__SharedDefs.Blake2_generic

type bytes

val hash : ?⁠key:bytes -> bytes -> int -> bytes

hash ?key msg size hashes msg and returns a digest of length size. An optional key argument can be passed for keyed hashing.

module Noalloc : sig ... end

Version of this function which writes its output in a buffer passed in as an argument