Module type Hacl_star__SharedDefs.ECDSA

type bytes = CBytes.t

val sign : sk:bytes -> msg:bytes -> k:bytes -> bytes option

sign sk msg k attempts to sign the message msg with secret key sk and signing secret k and returns the signature if successful.

val verify : pk:bytes -> msg:bytes -> signature:bytes -> bool

verify pk msg signature checks the signature of msg using public key pk and returns true if it is valid.

module Noalloc : sig ... end

Versions of these functions which write their output in a buffer passed in as an argument