Module type Hacl_star__SharedDefs.HKDF

type bytes = CBytes.t

val extract : salt:bytes -> ikm:bytes -> bytes

extract salt ikm computes a pseudorandom key using input key material ikm and salt salt.

val expand : prk:bytes -> info:bytes -> size:int -> bytes

expand prk info size expands the pseudorandom key prk, taking the info string info into account and returns a buffer of size bytes.

module Noalloc : sig ... end

Versions of these functions which write their output in a buffer passed in as an argument