Module HashFunction.Noalloc

Version of this function which writes its output in a buffer passed in as an argument

val hash : msg:bytes -> digest:bytes -> unit

hash msg digest hashes msg and outputs the result in digest.